añadida pestaña de exercicios

+++ public/php_functions/deleteAsignatura.php
@@ -6,18 +6,25 @@ if( !isset($_SESSION["rol"]) || ( $_SESSION["rol"] != "administrador" && $_SESSI
 require "../../vendor/autoload.php";
 
 use Clases\CursosDB;
+use Clases\ExercicioDB;
 
 $id = json_decode(file_get_contents("php://input"), true);
 
-$db = new CursosDB();
 
-if ( $db->deleteAsignatura($id) ) {
-    
-    $db->cerrarConexion();
-    $output = "<div class='alert alert-success'>Asignatura borrada correctamente</div>";    
+$db = new ExercicioDB();
+if ( empty( $db->getExercicioByAsignatura($id)) ) {
+
+    $db = new CursosDB();
+    if ( $db->deleteAsignatura($id) ) {    
+        $db->cerrarConexion();
+        $output = "<div class='alert alert-success'>Asignatura borrada correctamente</div>";    
+    }else {
+        $db->cerrarConexion();
+        $output = "<div class='alert alert-danger'>Non se pudo borrar a asignatura</div>";
+    }
 }else {
     $db->cerrarConexion();
-    $output = "<div class='alert alert-danger'>Non se pudo borrar a asignatura</div>";
+    $output = "<div class='alert alert-danger'>Non se pode eliminar a asignatura mentras existan exercicios desta asignatura</div>";
 }
 
 echo json_encode($output, JSON_UNESCAPED_UNICODE);
